+++ CHANGELOG.md
@@ -2,7 +2,7 @@ This is intended as an overview of the major changes
 
 v0.4-SNAPSHOT
 ===
-- Fix parsing a device name with a space in it.
+- Fix parsing a device name with a space ' ' or dot '.' in it.
 - Parse the list of power failures
 
 v0.3